+++ src/main/java/com/example/gak/global/redis/event/WaitingRoomUpdateEvent.java
@@ -1,4 +1,4 @@
-package com.example.gak.global.redis;
+package com.example.gak.global.redis.event;
 
 public record WaitingRoomUpdateEvent(Long sessionId) {
 }
