+++ src/main/java/com/example/gak/global/redis/eventHandler/WaitingRoomEventHandler.java
@@ -1,19 +1,22 @@
-package com.example.gak.global.redis;
+package com.example.gak.global.redis.eventHandler;
 
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.event.TransactionPhase;
 import org.springframework.transaction.event.TransactionalEventListener;
 
+import com.example.gak.global.redis.RedisPublisher;
+import com.example.gak.global.redis.event.WaitingRoomUpdateEvent;
+
 import lombok.RequiredArgsConstructor;
 
 @Component
 @RequiredArgsConstructor
-public class InProgressRoomEventHandler {
+public class WaitingRoomEventHandler {
 
 	private final RedisPublisher redisPublisher;
 
 	@TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
-	public void handle(InProgressRoomUpdateEvent event) {
-		redisPublisher.inProgressRoomPublish(event.sessionId());
+	public void handle(WaitingRoomUpdateEvent event) {
+		redisPublisher.waitingRoomPublish(event.sessionId());
 	}
 }
